New version with lots of changes
Mostly changes to the way the results are grabbed - before this plugin
worked by people having a preset list of MLS numbers and so the only
nightly process was an update on existing properties. Now that the
plugin is set to get all it will do an initial pull of 25 properties and
then add 25 new ones each night that were added the day before (I say
night but sometimes they'll be added during the day - each 24 hours from
when you first activate the plugin).